fix: open just the MJPEG if webrtc is not available
