+++ fury/stream/server/www/index_mjpeg.html
@@ -48,8 +48,6 @@
     <input id="use-stun" type="checkbox"/>
     <label for="use-stun">Use STUN server</label>
 </div> -->
-<button id="startBtn" class='hidden'>Start</button>
-<button id="stop" style="display: none" onclick="stop()">Stop</button>
 
 <div id="media">
 
@@ -64,8 +62,7 @@
     src='' id='videoMJPEG'/>
 </div>
 
-<script src="//webrtc.github.io/adapter/adapter-latest.js"></script>
-<script type='module' src="js/main.js">
+<script type='module' src="js/main_just_mjpeg.js">
 </script>
 </body>
 </html>
