+++ fury/stream/server/www/js/constants.js
@@ -1,5 +1,4 @@
 export const urlParams = new URLSearchParams(window.location.search);
-
 export const weelInterval =
   urlParams.get("weel_interval") === null
     ? 30
@@ -10,21 +9,6 @@ export const encoding =
     ? "webrtc"
     : urlParams.get("encoding");
 
-function getVideoDOMEL(encoding) {
-  let videoElId;
-  switch (encoding) {
-    case "webrtc":
-      videoElId = "video";
-      break;
-    case "mjpeg":
-      videoElId = "videoMJPEG";
-      break;
-    default:
-      videoElId = "video";
-  }
-  return document.getElementById(videoElId)
-}
-export const videoEl = getVideoDOMEL(encoding)
 export const mouseInterval =
   urlParams.get("mouse_interval") === null
     ? 80
@@ -37,4 +21,3 @@ const socketProtocol =
     : 'ws';
 
 export const socketAddr = `${socketProtocol}://${location.hostname}:${location.port}/ws`
-export const httpURL = `${location.protocol}://${location.hostname}:${location.port}/`
